week-1/requirements.txt
google-cloud-storage
argparse 